pyproject.toml
[project]
name = "mindmapper"
version = "0.1.0"
description = "MindMapper text-to-graph pipeline"
# readme = "README.md"  # Optional - uncomment if you add a README.md file
requires-python = ">=3.10"
dependencies = [
    "flask",
    "numpy",
    "scikit-learn",
    "nltk",
    "pyvis",
    "cohere",
    "gunicorn",
    "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["mindmapper"]